add Script, Config, Db
This is the first pass Script mixin.  It adds the same DSL/API
as the Dumpdb mixin with a few exceptions.

Unlike Dumpdb, this uses a config instance to store the DSL values.
This is to avoid messing with global state and to make the DSL
logic more easily tested.  Also, unlike Dumpdb, scripts demeter
their source/targets.  Finally, unlike Dumpdb, scripts track their
copy dump cmd *args*, not the full cmd strings.  This is b/c
we are going to let the (coming) Dk tasks choose how to run
the copy dump cmd based on the script's `ssh?` value.  This allows
reuse of all the Dk ssh logic and cmd building so we don't have to
reimplement it.

Also, unlike Dumpdb, there are no callbacks or run logic in
scripts.  Running and callbacks will be handled by Dk tasks.

This brings in the Db logic directly from Dumpdb - it is being
used almost exactly as-is.  The only difference as these db objs
know if they are equal to another db or not (for testing).

One other change from Dumpdb, the "settings" from Dumpdb have been
renamed as Config "values".  They are now organized under the
Config obj but this is just a naming organization change.  They
function just like in Dumpdb with the exception of the copy dump
cmd args value which is mentioned above.

This is the first part of porting Dumpdb logic here.  In a coming
effort, I'll add the task mixin for defining tasks to run scripts.