basic task mixin and internal task framework
This sets up the basic task mixin and the framework for the
internal tasks that script tasks will run.  I'm choosing to model
each "phase" of a script as a separate task so that user's can
bind callbacks in their dk configure blocks.  This removes the
need for dk-dumpdb to manage callbacks and allows the user to just
use what Dk makes available.

This doesn't actually implement any of the tasks.  The main mixin
`run!` method and the setup/teardown tasks have some pseudo-code
commented out for reference.  In a coming effort, I'll implement
everything.

This does go head and implement/test an InternalTask mixin.  This
is used by the internal tasks to provide some common logic:
specifically the logic for whether to run source cmds using ssh
or not.  This allows the tasks to just concentrate on whether
their cmds affect the source of the target and not on the
repetetive details of managing source cmds based on the script.