implement the Dump, CopyDump and Restore tasks
These tasks run the dump cmds, copy the dump from source to target
and run the restore cmds respectively.  This also updates the
InternalTask helper to define a `copy_cmd!` method.  This method,
like `source_cmd!`, looks at the script's ssh settings and then
builds the appropriate copy cmd based on them.  This is used by
the CopyDump task to implement the copy.

All of this logic is copied form the Runner logic in dumpdb and is
part of bringing in the dumpdb logic here so it can be run using
Dk.