auto-create a Task for each Script
This updates the script mixin to, when included on a receiver,
create a Task subclass for the receiver and include the Task mixin
and set the receiver as the script class.  This does the crappy
boilerplate work on behalf of the user so they have tasks ready to
run for their scripts.