switch to using scp for the copy dump file cmd
This is apparently faster than sftp but less feature-ful.  We
don't need a bunch of features for the copy so switch to get some
extra speed in copying.

This also switches off of using Dk's configured ssh args.  This
was a bug with the previous sftp implementation and is also an
issue with this scp implementation. The ssh args and scp (or sftp)
args aren't common so we can't blindly rely on them.  Really the
only "common" args are the `-o *` options. This switches to looking
for optional args in a named param so that the scp args can be
distinct from the ssh args.